pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "r2rome"
version = "0.1.0"
description = "Visualize nested project complexity as navigable digraphs"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
  { name = "Phil Sakievich" },
]
keywords = ["graphviz", "project-management", "digraph", "visualization", "planning"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Documentation",
  "Topic :: Scientific/Engineering :: Visualization",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "pyyaml>=6.0",
  "graphviz>=0.20",
  "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
]

[project.scripts]
r2rome = "r2rome.cli:main"

[project.urls]
Homepage = "https://github.com/psakievich/r2rome"
Issues = "https://github.com/psakievich/r2rome/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/r2rome"]

# Ensure non-Python template files are included in the wheel
[tool.hatch.build.targets.wheel.shared-data]

[tool.pytest.ini_options]
testpaths = ["tests"]
# addopts = "--cov=r2rome --cov-report=term-missing"  # requires pytest-cov; run manually with: pip install ".[dev]"

[tool.coverage.run]
source = ["src/r2rome"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
]
